# packages/shared/openapi.json
{
  "components": {
    "schemas": {
...
      "AudioAssetPublic": {
        "properties": {
          "accent": {
            "title": "Accent",
            "type": "string"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "voice": {
            "title": "Voice",
            "type": "string"
          }
        },
        "required": [
          "id",
          "storage_key",
          "url",
          "duration_ms",
          "accent",
          "voice"
        ],
        "title": "AudioAssetPublic",
        "type": "object"
      },
...
      "AudioConfirm": {
        "description": "Bước 4 cho audio tải lên (ADR-006 §2.3).\n\n`duration_ms` do TRÌNH DUYỆT khai, và đó là ngoại lệ có chủ ý với luật \"không\ntin lời trình duyệt\". Luật đó bảo vệ những thứ quyết định *tính đúng đắn* —\nfile có tồn tại không, đúng định dạng không, bao nhiêu byte — và ba thứ đó\nvẫn được hỏi lại kho lưu trữ. Độ dài clip thì chỉ dùng để vẽ thanh phát; đo\nnó ở máy chủ cần đọc khung mp3, mà thư viện làm việc đó nằm sau extra\n`content` — tức là đúng thứ tiến trình HTTP không được import (A4.1).",
        "properties": {
          "accent": {
            "title": "Accent",
            "type": "string"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          },
          "voice": {
            "default": "uploaded",
            "title": "Voice",
            "type": "string"
          }
        },
        "required": [
          "storage_key",
          "duration_ms",
          "accent"
        ],
        "title": "AudioConfirm",
        "type": "object"
      },
      "AudioRequestAck": {
        "description": "Chuông đã kêu hay chưa — KHÔNG phải audio đã có hay chưa.\n\n`queued=False` nghĩa là Redis không với tới được, không phải là thất bại:\nvòng quét định kỳ của worker vẫn tìm ra cùng số việc đó, chỉ muộn hơn.",
        "properties": {
          "queued": {
            "title": "Queued",
            "type": "boolean"
          }
        },
        "required": [
          "queued"
        ],
        "title": "AudioRequestAck",
        "type": "object"
      },
...
      "AudioTicketRequest": {
        "properties": {
          "ext": {
            "default": "mp3",
            "enum": [
              "mp3",
              "m4a",
              "wav"
            ],
            "title": "Ext",
            "type": "string"
          }
        },
        "title": "AudioTicketRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImageAssetPublic": {
        "properties": {
          "alt_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Text"
          },
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "license": {
            "title": "License",
            "type": "string"
          },
          "mime_type": {
            "title": "Mime Type",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "storage_key",
          "url",
          "mime_type",
          "size_bytes",
          "width",
          "height",
          "source",
          "source_url",
          "license",
          "attribution",
          "alt_text"
        ],
        "title": "ImageAssetPublic",
        "type": "object"
      },
      "ImageConfirm": {
        "description": "Bước 4: báo đã tải xong, kèm thông tin bản quyền.\n\nBa trường bản quyền là **bắt buộc**, khớp với ba cột NOT NULL trên\n`image_asset`. ADR-004 §2 nêu lý do: phần lớn ảnh mở là CC-BY — được dùng\n*với điều kiện* ghi công — và thông tin đó chỉ ghi lại trung thực được vào\nđúng lúc người ta thêm ảnh, khi trang nguồn còn đang mở.\n\nVới ảnh tự chụp thì `source_url` là nơi giữ bản gốc và `license` ghi rõ là\ncủa mình; trường này không phải thủ tục, nó là câu trả lời cho \"ai được phép\ndùng ảnh này\".",
        "properties": {
          "alt_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alt Text"
          },
          "attribution": {
            "minLength": 1,
            "title": "Attribution",
            "type": "string"
          },
          "license": {
            "maxLength": 64,
            "minLength": 1,
            "title": "License",
            "type": "string"
          },
          "source_url": {
            "maxLength": 1024,
            "minLength": 1,
            "title": "Source Url",
            "type": "string"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          }
        },
        "required": [
          "storage_key",
          "source_url",
          "license",
          "attribution"
        ],
        "title": "ImageConfirm",
        "type": "object"
      },
...
      "UploadTicket": {
        "description": "Vé để trình duyệt tự tải file lên, không đi qua API (ADR-006 §2.3).",
        "properties": {
          "allowed_formats": {
            "items": {
              "type": "string"
            },
            "title": "Allowed Formats",
            "type": "array"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "integer"
          },
          "fields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Fields",
            "type": "object"
          },
          "max_bytes": {
            "title": "Max Bytes",
            "type": "integer"
          },
          "method": {
            "default": "POST",
            "enum": [
              "POST",
              "PUT"
            ],
            "title": "Method",
            "type": "string"
          },
          "storage_key": {
            "title": "Storage Key",
            "type": "string"
          },
          "upload_url": {
            "title": "Upload Url",
            "type": "string"
          }
        },
        "required": [
          "upload_url",
          "fields",
          "storage_key",
          "max_bytes",
          "allowed_formats",
          "expires_at"
        ],
        "title": "UploadTicket",
        "type": "object"
      },
      "UploadTicketRequest": {
        "properties": {
          "ext": {
            "default": "jpg",
            "enum": [
              "jpg",
              "jpeg",
              "png",
              "webp"
            ],
            "title": "Ext",
            "type": "string"
          }
        },
        "title": "UploadTicketRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TOEIC Pilot API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/media/audio/confirm": {
      "post": {
        "operationId": "audio_confirm_api_v1_admin_media_audio_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioAssetPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audio Confirm",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/admin/media/audio/requests": {
      "post": {
        "description": "Đánh chuông cho worker TTS. **Không** sinh audio ở đây (ADR-007 §2.7b).\n\n202 chứ không 200, và đó là điều duy nhất cần hiểu về endpoint này: nó\nKHÔNG hứa audio đã có. API không sinh audio được — không import được\n`app.content` (A4.1), ảnh production không có edge-tts lẫn ffmpeg.\n\nKhông ghi bảng nào. Hàng đợi vẫn là *câu hỏi* \"nội dung nào thiếu audio\",\nnên bấm mười lần cũng không tạo ra mười job — worker chỉ hỏi lại database\nsớm hơn.\n\nChuông không kêu (Redis chết) vẫn trả 202, chỉ khác `queued`. Vòng quét định\nkỳ vẫn bắt được việc; nói dối rằng thất bại sẽ khiến biên tập viên bấm lại\nmột thứ vốn đã sẽ chạy.",
        "operationId": "request_audio_api_v1_admin_media_audio_requests_post",
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioRequestAck"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request Audio",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/admin/media/audio/ticket": {
      "post": {
        "description": "Vé upload cho audio. Khoá do phía ta sinh, y như đường ảnh.",
        "operationId": "audio_ticket_api_v1_admin_media_audio_ticket_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AudioTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTicket"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audio Ticket",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/admin/media/images": {
      "get": {
        "description": "Thư viện ảnh đã tải lên, mới nhất trước.\n\nKhông lọc theo `source`: ảnh lấy từ kho CC (`sourced`) và ảnh tự đưa lên\n(`uploaded`) cùng là ảnh dùng được cho một câu hỏi, và tách hai danh sách sẽ\nkhiến biên tập viên phải nhớ mình đã thêm ảnh nào bằng đường nào.",
        "operationId": "list_images_api_v1_admin_media_images_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ImageAssetPublic"
                  },
                  "title": "Response List Images Api V1 Admin Media Images Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Images",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/admin/media/images/confirm": {
      "post": {
        "operationId": "image_confirm_api_v1_admin_media_images_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImageConfirm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImageAssetPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Image Confirm",
        "tags": [
          "media"
        ]
      }
    },
    "/api/v1/admin/media/images/ticket": {
      "post": {
        "description": "Cấp một vé upload đã ký.\n\nKhoá do PHÍA TA sinh, từ một id ngẫu nhiên — client không được chọn nơi file\nsẽ nằm. Để client đặt khoá thì một người có vé hợp lệ có thể ghi đè lên\nđường dẫn của người khác, và chữ ký khi đó chỉ chứng minh \"ai đó được phép\nupload\", không chứng minh \"được phép upload vào đúng chỗ này\".",
        "operationId": "image_ticket_api_v1_admin_media_images_ticket_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTicket"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Image Ticket",
        "tags": [
          "media"
        ]
      }
    },
...
  }
}